1. Advertising/PRDs/Feature Automated KWs Research/KW-Classification/requirements.txt
# Braintrust SDK
braintrust>=0.0.100

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Data handling
pydantic>=2.0.0
jsonlines>=4.0.0

# Environment
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.25.0
